+++ app/components/IdentityCard/IdentityCard.tsx
@@ -25,6 +25,7 @@ import Form from '../Form';
 import Button from '~/cross-app-components/Button';
 import { useUpdateEffect } from '~/utils/hooks';
 import { editIdentityName } from '~/features/IdentitySlice';
+import DeleteIdentity from './DeleteIdentity';
 
 import styles from './IdentityCard.module.scss';
 
@@ -49,6 +50,7 @@ function statusImage(status: IdentityStatus) {
     switch (status) {
         case IdentityStatus.Confirmed:
             return <SuccessImage />;
+        case IdentityStatus.RejectedAndWarned:
         case IdentityStatus.Rejected:
             return <RejectedImage />;
         case IdentityStatus.Pending:
@@ -109,7 +111,12 @@ function IdentityListElement({
                     ) : null}
                     {statusImage(identity.status)}
                     <span className={clsx(styles.rightAligned, 'body2')}>
-                        Identity
+                        {identity.status === IdentityStatus.RejectedAndWarned &&
+                        canEditName ? (
+                            <DeleteIdentity identity={identity} />
+                        ) : (
+                            'Identity'
+                        )}
                     </span>
                 </div>
                 <Form<EditIdentityForm>