+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 -   Added an option to add an address book entry while creating a transfer transaction.
 -   Added an introductory screen to set up a node connection for first time users.
 -   Now shows connected node status in side bar.
+-   It is now possible to remove a failed identity.
 
 ### Changed